requirements.txt
openai>=1.53.0
click>=8.1.7
rich>=13.7.0
python-docx>=1.1.2
python-dotenv>=1.0.1
mutagen>=1.47.0
pydub>=0.25.1
pyannote.audio>=3.1.1
torch>=2.1.0
google-generativeai>=0.3.0


